requirements.txt
Pillow==8.2.0

httpx==0.17.1